Move release to commands.
